+++ AdvancedC#/FileSystemVisitorPL/FileSystemVisitor.cs
@@ -1,19 +1,5 @@
 ﻿namespace FileSystemVisitorPL
 {
-    public class FileSystemEventArgs : EventArgs
-    {
-        public bool AbortSearch { get; set; }
-
-        public bool ExcludeItem { get; set; }
-
-        public string Item { get; set; }
-
-        public FileSystemEventArgs(string item)
-        {
-            Item = item;
-        }
-    }
-
     public class FileSystemVisitor : IFileSystemVisitor
     {
         public event EventHandler SearchStarted;
